Consolidate type guards to types/entity/guards.ts
