+++ src/lib/types/entity/index.ts
@@ -1,9 +1,9 @@
 export const ROLES_LIST = ["admin", "superadmin"] as const;
 export type Role = (typeof ROLES_LIST)[number];
+
 /**
  * Sessions without tokens.
  */
-
 export interface SafeSession {
 	id: string;
 	userId: string;
